+++ src/main/kotlin/icu/windea/pls/lang/editor/folding/ParadoxVariableOperationExpressionFoldingBuilder.kt
@@ -4,6 +4,7 @@ import com.intellij.lang.*
 import com.intellij.lang.folding.*
 import com.intellij.openapi.editor.*
 import com.intellij.psi.*
+import icu.windea.pls.lang.*
 
 class ParadoxVariableOperationExpressionFoldingBuilder : ParadoxExpressionFoldingBuilder() {
     object Constants {
@@ -24,11 +25,11 @@ class ParadoxVariableOperationExpressionFoldingBuilder : ParadoxExpressionFoldin
     }
 
     override fun isCollapsedByDefault(node: ASTNode): Boolean {
-        return ParadoxFoldingSettings.getInstance().variableOperationExpressions
+        return getSettings().folding.variableOperationExpressionsByDefault
     }
 
     override fun buildFoldRegions(root: PsiElement, document: Document, quick: Boolean): Array<FoldingDescriptor> {
-        if (!ParadoxFoldingSettings.getInstance().variableOperationExpressionsEnabled) return FoldingDescriptor.EMPTY_ARRAY
+        if (!getSettings().folding.variableOperationExpressions) return FoldingDescriptor.EMPTY_ARRAY
 
         return super.buildFoldRegions(root, document, quick)
     }