+++ src/main/kotlin/icu/windea/pls/localisation/editor/folding/ParadoxLocalisationFoldingBuilder.kt
@@ -7,7 +7,9 @@ import com.intellij.openapi.project.*
 import com.intellij.openapi.util.*
 import com.intellij.psi.*
 import icu.windea.pls.*
+import icu.windea.pls.lang.*
 import icu.windea.pls.lang.editor.folding.*
+import icu.windea.pls.lang.settings.*
 import icu.windea.pls.localisation.psi.*
 import icu.windea.pls.localisation.psi.ParadoxLocalisationElementTypes.*
 
@@ -23,54 +25,55 @@ class ParadoxLocalisationFoldingBuilder : CustomFoldingBuilder(), DumbAware {
     }
 
     override fun isRegionCollapsedByDefault(node: ASTNode): Boolean {
+        val settings = getSettings().folding
         return when (node.elementType) {
-            COMMENT -> ParadoxFoldingSettings.getInstance().comment
-            PROPERTY_REFERENCE -> ParadoxFoldingSettings.getInstance().localisationReferencesFully
-            ICON -> ParadoxFoldingSettings.getInstance().localisationIconsFully
+            COMMENT -> settings.commentByDefault
+            PROPERTY_REFERENCE -> settings.localisationReferencesFullyByDefault
+            ICON -> settings.localisationIconsFullyByDefault
             COMMAND -> {
                 val conceptNode = node.findChildByType(CONCEPT)
                 if (conceptNode == null) {
-                    ParadoxFoldingSettings.getInstance().localisationCommands
+                    settings.localisationCommandsByDefault
                 } else {
-                    ParadoxFoldingSettings.getInstance().localisationConcepts
+                    settings.localisationConceptsByDefault
                 }
             }
-            CONCEPT_TEXT -> ParadoxFoldingSettings.getInstance().localisationConceptTexts
+            CONCEPT_TEXT -> settings.localisationConceptTextsByDefault
             else -> false
         }
     }
 
     override fun buildLanguageFoldRegions(descriptors: MutableList<FoldingDescriptor>, root: PsiElement, document: Document, quick: Boolean) {
-        val settings = ParadoxFoldingSettings.getInstance()
+        val settings = getSettings().folding
         collectDescriptorsRecursively(root.node, document, descriptors, settings)
     }
 
-    private fun collectDescriptorsRecursively(node: ASTNode, document: Document, descriptors: MutableList<FoldingDescriptor>, settings: ParadoxFoldingSettings) {
+    private fun collectDescriptorsRecursively(node: ASTNode, document: Document, descriptors: MutableList<FoldingDescriptor>, settings: ParadoxSettingsState.FoldingState) {
         when (node.elementType) {
             COMMENT -> {
-                if (settings.commentEnabled) {
+                if (settings.comment) {
                     ParadoxFoldingManager.addCommentFoldingDescriptor(node, document, descriptors)
                 }
             }
             LOCALE -> return //optimization
             PROPERTY_REFERENCE -> {
-                if (settings.localisationReferencesFullyEnabled) {
+                if (settings.localisationReferencesFully) {
                     descriptors.add(FoldingDescriptor(node, node.textRange))
                 }
             }
             ICON -> {
-                if (settings.localisationIconsFullyEnabled) {
+                if (settings.localisationIconsFully) {
                     descriptors.add(FoldingDescriptor(node, node.textRange))
                 }
             }
             COMMAND -> {
                 val conceptNode = node.findChildByType(CONCEPT)
                 if (conceptNode == null) {
-                    if (settings.localisationCommandsEnabled) {
+                    if (settings.localisationCommands) {
                         descriptors.add(FoldingDescriptor(node, node.textRange, null, PlsConstants.Strings.commandFolder))
                     }
                 } else {
-                    if (settings.localisationConceptsEnabled) {
+                    if (settings.localisationConcepts) {
                         val conceptTextNode = conceptNode.findChildByType(CONCEPT_TEXT)
                         if (conceptTextNode == null) {
                             descriptors.add(FoldingDescriptor(node, node.textRange, null, PlsConstants.Strings.conceptFolder))
@@ -82,7 +85,7 @@ class ParadoxLocalisationFoldingBuilder : CustomFoldingBuilder(), DumbAware {
             }
             CONCEPT_NAME -> return //optimization
             CONCEPT_TEXT -> {
-                if (settings.localisationConceptTextsEnabled) {
+                if (settings.localisationConceptTexts) {
                     descriptors.add(FoldingDescriptor(node, node.textRange))
                 }
             }