+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 * [ ] #128 [VIC3/CK3] Add ck3-tiger validator support to Plugin
 * [X] #136 [VIC3/CK3] Problems with ?= operator
+* [ ] Merge plugin code folding settings into plugin settings page and file / 将插件的代码折叠设置合并到插件的设置页面与文件
+* [X] 其他优化与BUG修复 / Other optimizations and bug fixes
 
 ## 1.3.32